requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
reportlab==4.1.0